+++ CHANGELOG.md
@@ -6,7 +6,10 @@ Version 2.1.0 *(2016-xx-xx)*
 * Improved: Scheduled transactions now have more accurate timestamps
 * Improved: Generate all scheduled transactions even if a scheduled is missed (e.g. device off)
 
-Version 2.0.6 *(2016-02-20)*
+Version 2.0.7 *(2016-04-20)*
+----------------------------
+* Fixed: Currency exchange rate does not accept very small rates (> 2 decimal places)
+* Improved: Updated translations for Japanese, Polish, French, Version 2.0.6 *(2016-02-20)*
 ----------------------------
 * Fixed: Saving transaction gets slower with increase in size of database
 * Fixed: Imbalance amount wrongly computed in split editor (for some accounts)